pyproject.toml
[project]
name = "medical-dictation"
version = "0.1.0"
description = "German medical dictation transcription and structured clinical summary extraction"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai-whisper>=20231117",
    "groq>=0.11.0",
    "click>=8.1.7",
    "rich>=13.7.0",
    "python-dotenv>=1.0.0",
    "torch>=2.1.0",
    "numpy>=1.24.0",
    "ffmpeg-python>=0.2.0",
]

[project.scripts]
medical-dictation = "medical_dictation.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/medical_dictation"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.3.0",
]
